Remove repo-dead selector helper and obsolete tests stub
