+++ box_management/builders/deposit_payloads.py
@@ -1,0 +1,15 @@
+from box_management.domain.legacy_utils import (
+    _build_comments_context_for_deposits,
+    _build_deposits_payload,
+    _build_reactions_from_instance,
+    _build_song_from_instance,
+    _build_user_from_instance,
+)
+
+__all__ = [
+    "_build_comments_context_for_deposits",
+    "_build_deposits_payload",
+    "_build_reactions_from_instance",
+    "_build_song_from_instance",
+    "_build_user_from_instance",
+]
