+++ box_management/domain/constants.py
@@ -1,0 +1,25 @@
+COMMENT_COOLDOWN_SECONDS = 120
+COMMENT_TARGET_USER_DAILY_LIMIT = 2
+
+COMMENT_REASON_ALREADY_COMMENTED = "already_commented"
+COMMENT_REASON_TARGET_USER_DAILY_COMMENT_LIMIT_REACHED = "target_user_daily_comment_limit_reached"
+COMMENT_REASON_RATE_LIMIT = "rate_limit"
+COMMENT_REASON_LINK_FORBIDDEN = "link_forbidden"
+COMMENT_REASON_EMAIL_FORBIDDEN = "email_forbidden"
+COMMENT_REASON_PHONE_FORBIDDEN = "phone_forbidden"
+COMMENT_REASON_EMPTY = "empty"
+COMMENT_REASON_TOO_LONG = "too_long"
+COMMENT_REASON_RESTRICTED = "restricted"
+COMMENT_REASON_REPORT_THRESHOLD = "report_threshold"
+COMMENT_REASON_RISK_QUARANTINE = "risk_quarantine"
+COMMENT_REASON_SPAM = "spam"
+COMMENT_REASON_HARASSMENT = "harassment"
+COMMENT_REASON_DOXXING = "doxxing"
+COMMENT_REASON_DELETE_BY_AUTHOR = "deleted_by_author"
+COMMENT_REASON_REMOVE_BY_MODERATION = "removed_by_moderation"
+
+COMMENT_REPORT_REASON_CHOICES = {
+    COMMENT_REASON_SPAM,
+    COMMENT_REASON_HARASSMENT,
+    COMMENT_REASON_DOXXING,
+}
