+++ box_management/services/boxes/incitations.py
@@ -1,0 +1,3 @@
+from box_management.domain.legacy_utils import _get_current_incitation_for_box
+
+__all__ = ["_get_current_incitation_for_box"]
