+++ box_management/services/comments/moderation_rules.py
@@ -1,0 +1,17 @@
+from box_management.domain.legacy_utils import (
+    _detect_comment_pre_creation_error,
+    _get_active_comment_restrictions_for_clients,
+    _get_profile_picture_url,
+    _log_blocked_comment_attempt,
+    _normalize_comment_text,
+    _score_comment_risk,
+)
+
+__all__ = [
+    "_detect_comment_pre_creation_error",
+    "_get_active_comment_restrictions_for_clients",
+    "_get_profile_picture_url",
+    "_log_blocked_comment_attempt",
+    "_normalize_comment_text",
+    "_score_comment_risk",
+]
