+++ box_management/services/deposits/accent_color.py
@@ -1,0 +1,3 @@
+from box_management.domain.legacy_utils import refresh_song_accent_color
+
+__all__ = ["refresh_song_accent_color"]
